refactor: web 모듈로 클래스 이동 및 패키지 정리
