+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,9 +1,10 @@
 package com.techcourse;
 
+import com.interface21.web.boot.WebApplicationInitializer;
+import com.interface21.webmvc.servlet.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.interface21.web.WebApplicationInitializer;
 
 /**
  * Base class for {@link WebApplicationInitializer}
