+++ app/src/main/java/com/techcourse/controller/annotationController/LoginViewController.java
@@ -3,8 +3,8 @@ package com.techcourse.controller.annotationController;
 import com.interface21.context.stereotype.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
 import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.view.JspView;
+import com.interface21.webmvc.view.ModelAndView;
+import com.interface21.webmvc.view.JspView;
 import com.techcourse.controller.common.UserSession;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
