+++ mvc/src/main/java/com/interface21/webmvc/exception/UncheckedServletException.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.exception;
 
 public class UncheckedServletException extends RuntimeException {
