+++ mvc/src/main/java/com/interface21/webmvc/handleradapter/HandlerAdapter.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handleradapter;
+package com.interface21.webmvc.handleradapter;
 
-import com.interface21.webmvc.servlet.ModelAndView;
+import com.interface21.webmvc.view.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
