+++ mvc/src/main/java/com/interface21/webmvc/handlermapping/ControllerScanner.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handlermapping;
+package com.interface21.webmvc.handlermapping;
 
 import com.interface21.context.stereotype.Controller;
 import org.reflections.Reflections;
