+++ mvc/src/main/java/com/interface21/webmvc/handlermapping/HandlerMappingRegistry.java
@@ -1,6 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.tobe.handlermapping;
+package com.interface21.webmvc.handlermapping;
 
-import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.annotation.HandlerMapping;
+import com.interface21.webmvc.handlermapping.annotation.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
 
 import java.util.ArrayList;
