+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -1,7 +1,7 @@
 package com.interface21.webmvc.servlet.mvc.tobe;
 
-import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.annotation.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.annotation.HandlerExecution;
+import com.interface21.webmvc.handlermapping.annotation.AnnotationHandlerMapping;
+import com.interface21.webmvc.handlermapping.annotation.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
